Changed four tables to numeric(18,4) for all share/unit columns:
- share_transactions.num_shares
- shareholders.num_shares
- stock_certificates.num_shares
- bills_of_sale.num_shares
Ensures decimal precision up to 4 places and no data loss.

X-Lovable-Edit-ID: edt-9aed45bc-bada-4a2d-92d6-ce016c9b44af

+++ src/components/AnnualMeetingWizard.tsx
@@ -745,7 +745,7 @@ export default function AnnualMeetingWizard({ company, onClose, onMeetingCreated
           return {
             meeting_id: mid,
             shareholder_name: m.name,
-            common_shares: m.units ? parseInt(m.units) : null,
+            common_shares: m.units ? parseFloat(m.units) : null,
             preferred_shares: m.interestPct ? parseFloat(m.interestPct) : null,
             distribution_amount: dist?.amount ? parseFloat(dist.amount.replace(/[,$]/g, "")) : null,
           };
